refactor: clean core mapping file
